dspy-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
dspy
python-dotenv==1.0.0
pydantic==2.9.0
httpx==0.27.0
openai==1.54.0
pytest==8.3.0
pytest-asyncio==0.25.0
packaging==24.2